requirements.txt
Django==1.10
djangorestframework==3.7.7
